fix applet setting styles
